requirements.txt
Flask
python-dotenv
requests
Pillow
pytesseract
opencv-python-headless
beautifulsoup4
google-genai